Reorganisation of the files: dir GameObjects: created in /Backend/GameObjects. file Armia: moved to GameObjects. file Plansza: -"- file Pole: -"- file Robal: -"- file Trader: -"-
dir GameMechanic: creted in /Backend/GameMechanic.
file GameMaster: moved to GameMechanic.
file Interfejs: moved to GameMechanic.
file InterfejsGracza: moved to GameMechanic.
file InterfejsNN: moved to GameMechanic.
